test: introduce canonical nexus testkit
