+++ tests/conftest.py
@@ -57,6 +57,8 @@ os.environ.setdefault("NEXUS_ENABLE_WRITE_BUFFER", "false")
 # ---------------------------------------------------------------------------
 os.environ.setdefault("NEXUS_ALLOW_EPHEMERAL_OAUTH_KEY", "1")
 
+from tests.testkit import make_test_nexus as make_test_nexus  # noqa: E402,F401
+
 # ---------------------------------------------------------------------------
 # Hypothesis profiles (Issue #1303)
 # ---------------------------------------------------------------------------
@@ -141,121 +143,6 @@ if _HAS_STRUCTLOG:
         structlog.contextvars.clear_contextvars()
         yield
         structlog.contextvars.clear_contextvars()
-
-
-def make_test_nexus(
-    tmp_path,
-    *,
-    backend=None,
-    permissions=None,
-    parsing=None,
-    cache=None,
-    memory=None,
-    distributed=None,
-    is_admin=False,
-    record_store=None,
-    use_raft=False,
-    metadata_store=None,
-    context=None,
-):
-    """Create a NexusFS instance for testing via factory (Issue #1801).
-
-    Uses ``create_nexus_fs()`` — the same boot path as production.
-    Defaults: permissions off, no auto-parse, no distributed features,
-    no bricks (kernel-only).
-
-    Args:
-        tmp_path: pytest tmp_path fixture for backend/metadata storage.
-        backend: Backend to mount at ``/``. Default: PathLocalBackend(tmp_path / "data").
-        permissions: PermissionConfig override. Default: enforce=False.
-        parsing: ParseConfig override. Default: auto_parse=False.
-        cache: CacheConfig override.
-        memory: MemoryConfig override.
-        distributed: DistributedConfig override. Default: all disabled.
-        is_admin: Admin flag for init_cred.
-        record_store: Optional RecordStoreABC.
-        use_raft: Use RaftMetadataStore (requires Python 3.13).
-        metadata_store: Override metadata store. Default: DictMetastore or Raft.
-        context: Override init_cred identity.
-
-    Returns:
-        NexusFS instance ready for testing.
-    """
-    from nexus.core.config import (
-        DistributedConfig,
-        ParseConfig,
-        PermissionConfig,
-    )
-    from nexus.factory import create_nexus_fs
-
-    if permissions is None:
-        permissions = PermissionConfig(enforce=False)
-    if parsing is None:
-        parsing = ParseConfig(auto_parse=False)
-    if distributed is None:
-        distributed = DistributedConfig(
-            enable_events=False,
-            enable_workflows=False,
-        )
-
-    if metadata_store is None:
-        # F2 C4 routes writes through ``kernel.sys_write``; DictMetastore is a
-        # Python-only dict so the kernel can't persist through it. Default to
-        # the same path production uses — RustMetastoreProxy wired to a fresh
-        # kernel + redb file under tmp_path. ``use_raft`` is now redundant
-        # but kept for API compatibility.
-        del use_raft
-        from nexus_runtime import PyKernel as _Kernel
-
-        from nexus.core.metastore import RustMetastoreProxy
-
-        _kernel = _Kernel()
-        # Phase 4 (full): drain federation's blob-fetcher slot + install
-        # the real `PeerBlobClient` so cross-node fetches work even
-        # inside test fixtures that exercise federation.
-        try:
-            import nexus_runtime as _nk
-
-            _nk.install_transport_wiring(_kernel)
-            _nk.install_federation_wiring(_kernel)
-        except Exception as _wiring_exc:
-            import logging as _logging
-
-            _logging.getLogger(__name__).debug(
-                "install_transport_wiring/install_federation_wiring failed in test fixture: %s",
-                _wiring_exc,
-            )
-        metadata_store = RustMetastoreProxy(_kernel, str(tmp_path / "metastore.redb"))
-
-    if backend is None:
-        from pathlib import Path
-
-        from nexus.backends.storage.path_local import PathLocalBackend
-
-        data_dir = Path(tmp_path) / "data"
-        data_dir.mkdir(exist_ok=True)
-        backend = PathLocalBackend(root_path=str(data_dir))
-
-    # Issue #1801: unified boot path — all NexusFS goes through factory
-    from tests.helpers.test_context import TEST_ADMIN_CONTEXT, TEST_CONTEXT
-
-    _init_cred = (
-        context if context is not None else (TEST_ADMIN_CONTEXT if is_admin else TEST_CONTEXT)
-    )
-
-    return create_nexus_fs(
-        backend=backend,
-        metadata_store=metadata_store,
-        record_store=record_store,
-        permissions=permissions,
-        parsing=parsing,
-        cache=cache,
-        memory=memory,
-        distributed=distributed,
-        is_admin=is_admin,
-        enabled_bricks=frozenset(),  # kernel-only for fast tests
-        init_cred=_init_cred,
-    )
 
 
 @pytest.fixture(autouse=True)
